// src/utils/types.ts
export type CartProduct = {
  id: string;
  title: string;
  subTitle?: string;
  image?: string;
  price?: number;
  mrp?: number;
};

export type CartItem = {
  product: CartProduct;
  quantity: number;
  totalPrice: number;
  billPrice: number;
  discount: number;
};